+++ lib/base_url_switcher.dart
@@ -3,9 +3,11 @@ library base_url_switcher;
 // Export all public APIs
 export 'models/environment.dart';
 export 'models/environment_type.dart';
+export 'models/configuration_item.dart';
 export 'services/env_service.dart';
 export 'services/base_url_manager.dart';
 export 'widgets/env_switcher.dart';
 export 'widgets/base_url_wrapper.dart';
 export 'widgets/simple_base_url_wrapper.dart';
+export 'widgets/configurable_base_url_wrapper.dart';
 export 'screens/env_switcher_screen.dart';